rename g3 rust module
