+++ pkm_rs/src/gen3_gba/pokemon_index.rs
